+++ src/repositories/metrics-repository-helper.js
@@ -1,3 +1,7 @@
+import { FormMetricName, FormMetricType } from '@defra/forms-model'
+
+import { getErrorMessage } from '~/src/helpers/error-message.js'
+import { logger } from '~/src/helpers/logging/logger.js'
 import { METRICS_COLLECTION_NAME, db } from '~/src/mongo.js'
 
 /**
@@ -27,7 +31,72 @@ export function getMetricCollection() {
 }
 
 /**
- * @import { Collection } from 'mongodb'
+ * Determines if any other publish events exist for this form
+ * @param {string} formId
+ * @param { string | undefined } language
+ * @param {ClientSession} session
+ * @returns {Promise<boolean>}
+ */
+export async function isFirstPublish(formId, language, session) {
+  const coll = getMetricCollection()
+
+  try {
+    const firstPublished = await coll.findOne(
+      {
+        type: FormMetricType.TimelineMetric,
+        metricName: FormMetricName.FormsFirstPublished,
+        formId,
+        ...getLanguageFilter(language)
+      },
+      { session }
+    )
+    return firstPublished === null
+  } catch (err) {
+    logger.error(
+      err,
+      `Failed to read timeline isFirstPublish for form id ${formId} - ${getErrorMessage(err)}`
+    )
+    throw err
+  }
+}
+
+/**
+ * Gets the earliest 'draft created' record of a form
+ * @param {string} formId
+ * @param { string | undefined } language
+ * @param {ClientSession} session
+ * @returns {Promise< WithId<FormTimelineMetric> | undefined >}
+ */
+export async function getFirstDraft(formId, language, session) {
+  const coll = getMetricCollection()
+
+  try {
+    const drafts = /** @type {WithId<FormTimelineMetric>[]} */ (
+      await coll
+        .find(
+          {
+            type: FormMetricType.TimelineMetric,
+            metricName: FormMetricName.NewFormsCreated,
+            formId,
+            ...getLanguageFilter(language)
+          },
+          { session }
+        )
+        .sort({ createdAt: 1 })
+        .toArray()
+    )
+    return drafts.length > 0 ? drafts[0] : undefined
+  } catch (err) {
+    logger.error(
+      err,
+      `Failed to read timeline getFirstDraft for form id ${formId} - ${getErrorMessage(err)}`
+    )
+    throw err
+  }
+}
+
+/**
+ * @import { ClientSession, Collection, WithId } from 'mongodb'
  * @import { FormOverviewMetric, FormTimelineMetric, FormTotalsMetric, FormDrilldownMetric } from '@defra/forms-model'
  * @import { FormMetricControl } from '~/src/service/metrics.js'
  */
